python101-assignment1-master-1671189243-main/requirements.txt
Django==3.2
djangorestframework==3.13.1
pytest==7.2.0
asgiref==3.5.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2